requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1      
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3  
attrs==23.1.0
azure-core==1.29.5    
azure-identity==1.15.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dataclasses-json==0.6.1
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.327
langsmith==0.0.56
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
msal==1.24.1
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
orjson==3.9.10
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.24.4
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
streamlit==1.28.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zipp==3.17.0